Spring Security 설정 및 JWT 설정 로그인 기능 구현

+++ src/main/java/com/bakery/dalcom/controller/UserController.java
@@ -20,7 +20,7 @@ public class UserController {
     }
 
     @GetMapping("/{userEid}")
-    public User getUserByUsername(@PathVariable int userEid) {
+    public User getUserByUsername(@PathVariable long userEid) {
         return userService.findByUserEid(userEid)
                 .orElseThrow(() -> new RuntimeException("User not found"));
     }
